clip/requirements.txt
torch
torchvision
torchaudio
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
Pillow
matplotlib
